+++ packages/react/src/checkbox/root/CheckboxRoot.test.tsx
@@ -557,6 +557,52 @@ describe('<Checkbox.Root />', () => {
       expect(button).to.have.attribute('aria-invalid', 'true');
     });
 
+    it('props: validationMode=onChange', async () => {
+      await render(
+        <Field.Root
+          validationMode="onChange"
+          validate={(value) => {
+            const checked = value as boolean;
+            return checked ? 'error' : null;
+          }}
+        >
+          <Checkbox.Root data-testid="button" />
+        </Field.Root>,
+      );
+
+      const button = screen.getByTestId('button');
+
+      expect(button).not.to.have.attribute('aria-invalid');
+
+      fireEvent.click(button);
+
+      expect(button).to.have.attribute('aria-invalid', 'true');
+    });
+
+    it('prop: validationMode=onBlur', async () => {
+      await render(
+        <Field.Root
+          validationMode="onBlur"
+          validate={(value) => {
+            const checked = value as boolean;
+            return checked ? 'error' : null;
+          }}
+        >
+          <Checkbox.Root data-testid="button" />
+          <Field.Error data-testid="error" />
+        </Field.Root>,
+      );
+
+      const button = screen.getByTestId('button');
+
+      expect(button).not.to.have.attribute('aria-invalid');
+
+      fireEvent.click(button);
+      fireEvent.blur(button);
+
+      expect(button).to.have.attribute('aria-invalid', 'true');
+    });
+
     it('Field.Label', async () => {
       const { container } = await render(
         <Field.Root>
