+++ packages/react/src/switch/root/SwitchRoot.test.tsx
@@ -465,6 +465,52 @@ describe('<Switch.Root />', () => {
       expect(button).to.have.attribute('aria-invalid', 'true');
     });
 
+    it('prop: validationMode=onChange', async () => {
+      await render(
+        <Field.Root
+          validationMode="onChange"
+          validate={(value) => {
+            const checked = value as boolean;
+            return checked ? 'error' : null;
+          }}
+        >
+          <Switch.Root data-testid="button" />
+        </Field.Root>,
+      );
+
+      const button = screen.getByTestId('button');
+
+      expect(button).not.to.have.attribute('aria-invalid');
+
+      fireEvent.click(button);
+
+      expect(button).to.have.attribute('aria-invalid', 'true');
+    });
+
+    it('prop: validationMode=onBlur', async () => {
+      await render(
+        <Field.Root
+          validationMode="onBlur"
+          validate={(value) => {
+            const checked = value as boolean;
+            return checked ? 'error' : null;
+          }}
+        >
+          <Switch.Root data-testid="button" />
+          <Field.Error data-testid="error" />
+        </Field.Root>,
+      );
+
+      const button = screen.getByTestId('button');
+
+      expect(button).not.to.have.attribute('aria-invalid');
+
+      fireEvent.click(button);
+      fireEvent.blur(button);
+
+      expect(button).to.have.attribute('aria-invalid', 'true');
+    });
+
     it('Field.Label', async () => {
       await render(
         <Field.Root>
